+++ DigitalSynthVRA8M/synth.h
@@ -24,12 +24,12 @@ public:
     if (is_data_byte(b)) {
       if (m_system_exclusive) {
         // do nothing
-      } else if (m_system_data_remaining != (uint8_t) 0) {
+      } else if (m_system_data_remaining != 0) {
         m_system_data_remaining--;
       } else if (m_running_status == (NOTE_ON | MIDI_CH)) {
         if (!is_data_byte(m_first_data)) {
           m_first_data = b;
-        } else if (b == (uint8_t) 0) {
+        } else if (b == 0) {
           note_off(m_first_data);
           m_first_data = DATA_BYTE_INVALID;
         } else {
